+++ policy_sentry/command/download_policies.py
@@ -1,5 +1,4 @@
 import click
-from pathlib import Path
 from policy_sentry.shared.download import download_remote_policies, download_policies_recursively
 from policy_sentry.shared.login import get_list_of_aws_profiles
 from policy_sentry.shared.constants import DEFAULT_CREDENTIALS_FILE