+++ policy_sentry/command/initialize.py
@@ -1,10 +1,7 @@
-#!/usr/bin/env python3
-
 import click
 from policy_sentry.shared.config import create_policy_sentry_config_directory, \
     create_audit_directory, create_default_overrides_file, create_policy_analysis_directory, \
     create_default_report_config_file
-from pathlib import Path
 from policy_sentry.shared.database import connect_db, create_database
 from policy_sentry.shared.file import get_list_of_service_prefixes_from_links_file
 from policy_sentry.shared.constants import HOME, CONFIG_DIRECTORY