+++ policy_sentry/command/write_policy.py
@@ -1,6 +1,5 @@
 import click
 import json
-from pathlib import Path
 from policy_sentry.shared.actions import get_all_actions, get_dependent_actions
 from policy_sentry.shared.database import connect_db
 from policy_sentry.shared.minimize import minimize_statement_actions